css change to add scrollbar
As part as adding class element to the filter body in Filter.js we add the css code here to add a max height to the Filter component and an overflow of auto. This change introduces the vertical scrollbar when the Filter component contains too much data, instead of expanding down the page.